src/functions/recovery/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.1
google-auth>=2.22.0
requests>=2.31.0
flask>=3.0.0
